+++ ChangeLog.md
@@ -13,12 +13,14 @@ Other enhancements:
 * Added a `stack config set resolver RESOLVER` command. Part of work on [#115](https://github.com/commercialhaskell/stack/issues/115)
 * `stack setup` can now install GHCJS on windows.  See [#1145](https://github.com/commercialhaskell/stack/issues/1145) and [#749](https://github.com/commercialhaskell/stack/issues/749)
 * `stack hpc report` command added, which generates reports for HPC tix files
+* `stack ghci` builds the project before launching GHCi.  If the build fails, optimistically launch GHCi anyway.  Use `stack ghci --no-build` option to disable [#1065](https://github.com/commercialhaskell/stack/issues/1065)
 
 Bug fixes:
 
-* Haddocks are now copied for dependencies [#1105](https://github.com/commercialhaskell/stack/issues/1105)
+* Fix: Haddocks not copied for dependencies [#1105](https://github.com/commercialhaskell/stack/issues/1105)
+* Fix: Global options did not work consistently after subcommand [#519](https://github.com/commercialhaskell/stack/issues/519)
+* Fix: 'stack ghci' doesn't notice that a module got deleted [#1180](https://github.com/commercialhaskell/stack/issues/1180)
 * Rebuild when cabal file is changed
-* Global options now work consistently after subcommand [#519](https://github.com/commercialhaskell/stack/issues/519)
 
 ## v0.1.6.0